requirements.txt
streamlit
groq
chromadb
python-dotenv
pypdf
python-docx